api-gateway/requirements.txt
django==5.1.5
psycopg2-binary==2.9.9
PyJWT==2.9.0
requests==2.32.3
